ml-service/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.1.0

# ML/AI
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
scikit-image==0.22.0
facenet-pytorch==2.5.3
mtcnn==0.1.1
insightface==0.7.3
deepface==0.0.79

# OCR
pytesseract==0.3.10
pillow==10.1.0
easyocr==1.7.1
pywt==1.4.1

# Image Processing
imutils==0.5.4
albumentations==1.3.1
imageio==2.33.0

# Redis & Async
redis==5.0.1
aioredis==2.0.1
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
cachetools==5.3.2
python-json-logger==2.0.7
prometheus-client==0.19.0

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0